Move checkpoint data selection to chainparams
