+++ src/chainparams.h
@@ -8,6 +8,7 @@
 
 #include "core.h"
 #include "chainparamsbase.h"
+#include "checkpoints.h"
 #include "protocol.h"
 #include "uint256.h"
 
@@ -77,6 +78,7 @@ public:
     const std::vector<CDNSSeedData>& DNSSeeds() const { return vSeeds; }
     const std::vector<unsigned char>& Base58Prefix(Base58Type type) const { return base58Prefixes[type]; }
     const std::vector<CAddress>& FixedSeeds() const { return vFixedSeeds; }
+    virtual const Checkpoints::CCheckpointData& Checkpoints() const = 0;
 protected:
     CChainParams() {}
